+++ cucumber-scala/src/test/scala/io/cucumber/scala/steps/errors/incorrectclasshooks/IncorrectClassHooksDefinition.scala
@@ -8,16 +8,22 @@ class IncorrectClassHooksDefinition extends ScalaDsl {
   // On a single line to avoid difference between Scala versions for the location
 
   // A body that does not return Unit => interpreted as missing body
+  BeforeAll { 22 }
+  
+  // A body that does not return Unit => interpreted as missing body
   Before { 1 }
 
   // A body that does not return Unit => interpreted as missing body
   BeforeStep { "toto" }
 
   // A body that does not return Unit => interpreted as missing body
+  AfterAll { 66 }
+  
+  // A body that does not return Unit => interpreted as missing body
   After { 33 }
 
   // A body that does not return Unit => interpreted as missing body
   AfterStep { "toto" }
-
+  
 }
 //@formatter:on