+++ cucumber-scala/src/test/scala/io/cucumber/scala/steps/objects/StepsInObject.scala
@@ -4,12 +4,22 @@ import io.cucumber.scala.{EN, ScalaDsl}
 
 object StepsInObject extends ScalaDsl with EN {
 
+  BeforeAll {
+    // Nothing
+    ()
+  }
+
   Before {
     // Nothing
     ()
   }
 
   BeforeStep {
+    // Nothing
+    ()
+  }
+
+  AfterAll {
     // Nothing
     ()
   }